Move tokenizer
