+++ algorithmic_efficiency/workloads/wmt/input_pipeline.py
@@ -6,7 +6,7 @@ from typing import Dict, List, Optional, Union
 import tensorflow as tf
 import tensorflow_datasets as tfds
 
-from algorithmic_efficiency.workloads.wmt.wmt_jax import tokenizer
+from algorithmic_efficiency.workloads.wmt import tokenizer
 
 AUTOTUNE = tf.data.AUTOTUNE
 Features = Dict[str, tf.Tensor]
@@ -304,11 +304,11 @@ def get_wmt_datasets(vocab_size: int,
   if vocab_path is None:
     vocab_path = os.path.expanduser('~/wmt_sentencepiece_model')
 
-  train_ds_builder = tfds.builder('wmt17_translate/de-en')
+  train_ds_builder = tfds.builder('wmt17_translate/de-en', data_dir='/mnt/qb/work/hennig/reschenhagen26/tensorflow_datasets/')
   train_data = get_raw_dataset(
       train_ds_builder, 'train', reverse_translation=reverse_translation)
 
-  eval_ds_builder = tfds.builder('wmt14_translate/de-en')
+  eval_ds_builder = tfds.builder('wmt14_translate/de-en', data_dir='/mnt/qb/work/hennig/reschenhagen26/tensorflow_datasets/')
   eval_data = get_raw_dataset(
       eval_ds_builder, 'test', reverse_translation=reverse_translation)
